# Conflicts:
#	SP4/Base.v12.suo
#	SP4/Base/Debug/Base.log
#	SP4/Common/Debug/Common.log

+++ SP4/Base/Source/Main/Game/Objects/Characters/Mage.cpp
@@ -50,11 +50,11 @@ void Mage::LevelUp(bool init)
 			return;
 	}
 
-	ExperiencePoints = ExperienceNeed[Level];
+	ExperiencePoints = (float)ExperienceNeed[Level];
 	SetHealth(MageHealth[Level]);
 	SetMaxHealth(MageHealth[Level]);
 	SetAttack(MageAttack[Level]);
-	SetDefence(MageDefence[Level]);
+	SetDefence((size_t)MageDefence[Level]);
 	SetDamageMitigation();
 	if (Level <= 10)
 	{
